Fix TODO items 426-441: M2Loader/SKELLoader anim loading fidelity & WMOLegacyLoader portal types
- 426-428: M2Loader::loadAnims() — add animIsChunked flag, skeletonBoneData/animData selection, _patch_bone_animation call
- 429-431: M2Loader::loadAnimsForIndex() — same fixes
- 432: Verified parseChunk_MD21_textures already matches JS
- 433,435: SKELLoader::loadAnims() — add skeletonBoneData/animData selection, _patch_bone_animation call
- 436-437: SKELLoader::loadAnimsForIndex() — same fixes
- 438: Verified double seek matches JS faithfully
- 439: Verified globalLoops type matches JS M2Loader
- 440: Verified triangleStart uint32_t is correct
- 441: Fixed ofsPortals/numPortals to uint32_t in WMOLegacyLoader

Agent-Logs-Url: https://github.com/NathanNatNat/wow.export.cpp/sessions/4a652020-417d-4c7d-83f7-d4cfcc875546

+++ src/js/3D/loaders/WMOLegacyLoader.cpp
@@ -530,8 +530,8 @@ void WMOLegacyLoader::parse_MOGP(BufferWrapper& data, uint32_t chunkSize) {
 
 	if (this->version == WMO_VER_ALPHA) {
 		// alpha uses uint32 for portal fields
-		this->ofsPortals = static_cast<uint16_t>(data.readUInt32LE());
-		this->numPortals = static_cast<uint16_t>(data.readUInt32LE());
+		this->ofsPortals = data.readUInt32LE();
+		this->numPortals = data.readUInt32LE();
 	} else {
 		this->ofsPortals = data.readUInt16LE();
 		this->numPortals = data.readUInt16LE();
